+++ src/GlobalSimulParameters.cpp
@@ -223,7 +223,7 @@ GSP::GSP(const string globalParamsFile)
 		}
 
 		m_FreqDist = GlobParms.get_val<int>("DIST_FREQ");
-		if (m_NoDist != m_FreqDist.size())
+		if (m_NoDist != (int)m_FreqDist.size())
 		{
 			logg.error("!!! Parameter DIST_FREQ : number of frequencies must be equal to the number of disturbances (DIST_NO)!");
 		}
@@ -266,7 +266,7 @@ GSP::GSP(const string globalParamsFile)
 		if (v_int.size()) m_NoFireDistSub = v_int[0]; else m_NoFireDistSub = 0;
 		v_int = GlobParms.get_val<int>("FIRE_FREQ",true);
 		if (v_int.size()) m_FreqFireDist = v_int; else m_FreqFireDist = vector<int>(1,0);
-		if (m_NoFireDist != m_FreqFireDist.size())
+		if (m_NoFireDist != (int)m_FreqFireDist.size())
 		{
 			logg.error("!!! Parameter FIRE_FREQ : number of frequencies must be equal to the number of fire disturbances (FIRE_NO)!");
 		}
@@ -364,7 +364,7 @@ GSP::GSP(const string globalParamsFile)
 	if (m_DoAliensIntroduction)
 	{
 		m_FreqAliens = GlobParms.get_val<int>("ALIENS_FREQ");
-		if (m_NoFG != m_FreqAliens.size())
+		if (m_NoFG != (int)m_FreqAliens.size())
 		{
 			logg.error("!!! Parameter ALIENS_FREQ : number of frequencies must be equal to the number of PFG (NO_PFG)!");
 		}
